--- koala/excel/__init__.py
@@ -1,2 +1,0 @@
-# import utils
-# import excel
