--- koala/openpyxl/__init__.py
@@ -1,4 +1,0 @@
-from __future__ import absolute_import
-# Copyright (c) 2010-2016 openpyxl
-
-from .tokenizer import Tokenizer
