+++ koala/openpyxl/cell/__init__.py
@@ -1,0 +1,9 @@
+#
+# source: https://bitbucket.org/openpyxl/openpyxl/src/93604327bce7aac5e8270674579af76d390e09c0/openpyxl/cell/__init__.py?at=default&fileviewer=file-view-default
+#_______________________________________________________________________________________________________________________________________________________________
+
+
+from __future__ import absolute_import
+# Copyright (c) 2010-2016 openpyxl
+
+# from .cell import *
