+++ koala/openpyxl/cell/text.py
@@ -1,7 +1,9 @@
+#
+# source: https://bitbucket.org/openpyxl/openpyxl/src/93604327bce7aac5e8270674579af76d390e09c0/openpyxl/formula/tokenizer.py?at=default&fileviewer=file-view-default
+#___________________________________________________________________________________________________________________________________________________________________
+
 from __future__ import absolute_import
 # Copyright (c) 2010-2016 openpyxl
-
-# source: https://bitbucket.org/openpyxl/openpyxl/src/93604327bce7aac5e8270674579af76d390e09c0/openpyxl/cell/text.py?at=default&fileviewer=file-view-default
 
 """
 Richtext definition
