+++ koala/openpyxl/utils/exceptions.py
@@ -1,3 +1,7 @@
+#
+# source: https://bitbucket.org/openpyxl/openpyxl/src/93604327bce7aac5e8270674579af76d390e09c0/openpyxl/utils/exceptions.py?at=default&fileviewer=file-view-default
+#__________________________________________________________________________________________________________________________________________________________________
+
 from __future__ import absolute_import
 # Copyright (c) 2010-2016 openpyxl
