+++ koala/openpyxl/xml/constants.py
@@ -1,3 +1,8 @@
+#
+# source: https://bitbucket.org/openpyxl/openpyxl/src/93604327bce7aac5e8270674579af76d390e09c0/openpyxl/xml/constants.py?at=default&fileviewer=file-view-default
+#_______________________________________________________________________________________________________________________________________________________________
+
+
 from __future__ import absolute_import
 
 """Constants for fixed paths in a file and xml namespace urls."""
