--- koala/xml/__init__.py
@@ -1,2 +1,0 @@
-import constants
-import functions
